+++ lib/elixir/src/elixir_compiler.erl
@@ -225,6 +225,7 @@ core_main() ->
     "lib/elixir/lib/kernel/cli.ex",
     "lib/elixir/lib/kernel/error_handler.ex",
     "lib/elixir/lib/kernel/parallel_compiler.ex",
+    "lib/elixir/lib/kernel/locals_tracker.ex",
     "lib/elixir/lib/kernel/record_rewriter.ex"
   ].