+++ lib/elixir/src/elixir_scope.erl
@@ -76,18 +76,19 @@ build_ex_var(Line, Key, Name, S) when is_integer(Line) ->
 
 % Handle Macro.Env conversion
 
-to_erl_env({ 'Elixir.Macro.Env', Module, File, _Line, Function, Aliases, Context, Requires, Functions, Macros, FileModules }) ->
+to_erl_env({ 'Elixir.Macro.Env', Module, File, _Line, Function, Aliases, Context,
+    Requires, Functions, Macros, FileModules, FunctionKind }) ->
   #elixir_scope{module=Module,file=File,
     function=Function,aliases=Aliases,context=Context,
     requires=Requires,macros=Macros,functions=Functions,
-    context_modules=FileModules}.
+    context_modules=FileModules,function_kind=FunctionKind}.
 
 to_ex_env({ Line, #elixir_scope{module=Module,file=File,
     function=Function,aliases=Aliases,context=Context,
     requires=Requires,macros=Macros,functions=Functions,
-    context_modules=FileModules} }) when is_integer(Line) ->
+    context_modules=FileModules,function_kind=FunctionKind} }) when is_integer(Line) ->
   { 'Elixir.Macro.Env', Module, File, Line, Function, Aliases,
-    Context, Requires, Functions, Macros, FileModules }.
+    Context, Requires, Functions, Macros, FileModules, FunctionKind }.
 
 % Provides a tuple with only the scope information we want to serialize.