+++ lib/elixir/src/elixir_translator.erl
@@ -578,7 +578,7 @@ translate_fn(Meta, Clauses, S) ->
 %% Locals
 
 translate_local(Meta, Name, Args, #elixir_scope{local=nil} = S) ->
-  elixir_import:record_local({ Name, length(Args) }, S),
+  elixir_def_local:record({ Name, length(Args) }, S),
   Line = ?line(Meta),
   { TArgs, NS } = translate_args(Args, S),
   { { call, Line, { atom, Line, Name }, TArgs }, NS };